code_new/requirements.txt
absl_py==2.0.0
numpy==2.4.4
openai==2.31.0
python-dotenv==1.2.2
scipy==1.17.1
sentence_transformers==5.4.0
